requirements.txt
asn1crypto==1.0.1
Babel==2.7.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
confusable-homoglyphs==3.2.0
cryptography==2.7
Django==2.2.6
django-crispy-forms==1.8.0
django-phonenumber-field==3.0.1
django-registration==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
idna==2.8
olefile==0.46
phonenumbers==8.10.23
Pillow==6.2.0
psycopg2==2.8.2
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
win-inet-pton==1.1.0
wincertstore==0.2
